ref(analysis): created a shared `utils` package
