+++ analysis/src/main/scala/org/polystat/odin/analysis/utils/inlining/LocateMethods.scala
@@ -1,10 +1,10 @@
-package org.polystat.odin.analysis.inlining
+package org.polystat.odin.analysis.utils.inlining
 
 import cats.data.{NonEmptyList => Nel}
 import cats.syntax.foldable._
 import higherkindness.droste.data.Fix
 import org.polystat.odin.analysis.ObjectName
-import org.polystat.odin.analysis.inlining.LocateCalls._
+import LocateCalls._
 import org.polystat.odin.core.ast._
 import org.polystat.odin.core.ast.astparams.EOExprOnly
