+++ analysis/src/test/scala/org/polystat/odin/analysis/inlining/InliningTests.scala
@@ -9,6 +9,7 @@ import InlineCallsTestCases._
 import cats.syntax.either._
 import cats.syntax.traverse._
 import cats.data.{EitherNel, NonEmptyList => Nel}
+import org.polystat.odin.analysis.utils.inlining.{Inliner, LocatorContext}
 // import org.polystat.odin.core.ast._
 
 class InliningTests extends AnyWordSpec {
@@ -27,7 +28,7 @@ class InliningTests extends AnyWordSpec {
         val obtained: EitherNel[String, String] = Parser
           .parse(before)
           .leftMap(Nel.one)
-          .flatMap(Context.setLocators)
+          .flatMap(LocatorContext.setLocators)
           .map(_.toEOPretty)
         assert(expected == obtained)
       }
